Simplified patch after OpenGLRaw's GL_FALSE/GL_TRUE change.

+++ src/Graphics/Luminance/Core/Shader/Program.hs
@@ -817,18 +817,18 @@ instance Uniform [V 4 Float] where
 -- Matrices --------------------------------------------------------------------
 instance Uniform (M44 Float) where
 #ifdef __GL45
-  toU prog l = pure . U $ \v -> with v $ glProgramUniformMatrix4fv prog l 1 (fromIntegral GL_FALSE) . castPtr
+  toU prog l = pure . U $ \v -> with v $ glProgramUniformMatrix4fv prog l 1 GL_FALSE . castPtr
 #elif defined(__GL33)
-  toU _ l = pure . U $ \v -> with v $ glUniformMatrix4fv l 1 (fromIntegral GL_FALSE) . castPtr
+  toU _ l = pure . U $ \v -> with v $ glUniformMatrix4fv l 1 GL_FALSE . castPtr
 #endif
 
 instance Uniform [M44 Float] where
 #ifdef __GL45
   toU prog l = pure . U $ \v -> withArrayLen v $ \size p ->
-    glProgramUniformMatrix4fv prog l (fromIntegral size) (fromIntegral GL_FALSE) (castPtr p)
+    glProgramUniformMatrix4fv prog l (fromIntegral size) GL_FALSE (castPtr p)
 #elif defined(__GL33)
   toU _ l = pure . U $ \v -> withArrayLen v $ \size p ->
-    glUniformMatrix4fv l (fromIntegral size) (fromIntegral GL_FALSE) (castPtr p)
+    glUniformMatrix4fv l (fromIntegral size) GL_FALSE (castPtr p)
 #endif
 
 --------------------------------------------------------------------------------
